// fe/src/types/generated/VersionPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Version payload containing the actual entity data
 */
export type VersionPayload = {
    version_number: number
    created_at: string
    created_by: string | null
    data: unknown
}